#51 [REFACTOR] : seperate profile auth and member
